services/order-service/requirements.txt
aiomysql==0.3.2
fastapi==0.127.0
firebase_admin==7.1.0
httpx==0.28.1
passlib==1.7.4
pydantic==2.12.5
pydantic_settings==2.12.0
PyJWT==2.10.1
python-jose==3.3.0
requests==2.32.5
SQLAlchemy==2.0.45
uvicorn==0.34.0
apscheduler==3.11.0
email-validator

python-barcode
pillow